requirements.txt
pandas==2.1.3
openpyxl==3.1.2
opencv-python==4.8.1.78
pyzbar==0.1.9
msoffcrypto-tool
flask==3.1.2
